Try to fix flaky test KillTaskIT#killTask()
The test was listing all tasks and then tried to kill the first task
from the list. Sometimes this was a short-living task that didn't exist
anymore when the delete was executed. In this case the response was 404
and not as expected 204, so that the test failed with:

killTaskTests_inOrder(com.google.gerrit.acceptance.rest.config.KillTaskIT)
java.lang.AssertionError: Expected status code 204: Not true that <404> is equal to <204>
  at com.google.gerrit.acceptance.RestResponse.assertStatus(RestResponse.java:45)
  at com.google.gerrit.acceptance.RestResponse.assertNoContent(RestResponse.java:65)
  at com.google.gerrit.acceptance.rest.config.KillTaskIT.killTask(KillTaskIT.java:37)
  at com.google.gerrit.acceptance.rest.config.KillTaskIT.killTaskTests_inOrder(KillTaskIT.java:60)
  at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
  ...

Instead of attempting to delete the first task from the list, always
delete the "Log File Compressor" task, which should be always present.

Change-Id: I46708710769a70eea81336e3c6a9ac8febf2b52a
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: e5be222059bec136de0b1eaafc41edb4b3783714