ChangeJson: Do not scan all starred changes refs of the change
When change details are loaded, the stars of the calling users are
included into the ChangeInfo. The stars are stored in
refs/starred-changes/<sharded-changeid>/<user-id> refs in the All-Users
repository and must be loaded from there. At the moment we load and
parse all starred changes ref of the change (all refs that start with
'refs/starred-changes/<sharded-changeid>/') when change details are
requested. Since we need only the stars of the calling user it is better
to only load and parse the single starred changes ref of the calling
user.

Change-Id: Ic218eaf2ce0e2b4d94c5c37136a86c20ee658dbd
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: df1b48976d11c1d5e632d196ad2ba2ebd7a16550