Add a User Preference to Receive only Plaintext Emails
This adds a user preference such that we only send plaintext emails to
this user. It also adds docs and tests.

For the user, this comes at the cost of not being in the same email
thread as the users who receive HTML emails. We decided to send out
plaintext emails on a one-by-one basis instead of making one thread (as
we do with HTML) as this would just create more confusion.

When admins enable inbound email for Gerrit, the aforementioned caveat
isn't a problem anymore as users would email Gerrit anyway instead of
using a notification email as messaging thread.

Change-Id: I3d0fc920c9d81e1a7b707ae5f3f4600a62b8a93b

Former-commit-id: 39777363a63c603cfa68ca52d2f3c99eb04f2f46