Better email wording when updating changes the user does not own
When a user updates a change that they are not the owner of (for
example, when providing assistance with a new patch set or when updating
the commit message on the owner's behalf) the generated email can give
the impression that the user has created an entirely new change.

With this change, ReplacePatchSet emails check whether the user who
modifies a change differs from the owner of that change. In this case it
appends a note regarding the owner.

For example, the sentence in the replace patch set emails ...

  User One has uploaded patch set #123 to this change.

... becomes ...

  User One has uploaded patch set #123 to the change originally created
  by User Two.

Bug: Issue 4805
Change-Id: Ia9aefaaf6fd3e5f0485675f30e048ac54c94a1b9

Former-commit-id: 2f7c13541654590299f5c3dcc556a6faa214c25c