Bazel: Fix eclipse project generation
The way currently eclipse classpath generation is working, sources
must be provided to generate source entry in .classpath for library
dependency. The dependency on "//gerrit-acceptance-tests:lib" is
included, but this library doesn't have any sources, that why it
is not included as source folder. To rectify it, add a dummy source
file.

Test Plan:

  $ ./tools/eclipse/project.py
  $ grep gerrit-acceptance-tests .classpath | wc -l
  1

Change-Id: Idc8534c9b94121f057da7727e3ffc3c04f83227f

Former-commit-id: 1c1e6ea1a0b5f4ad26aefab675bf4a4f361a7e8e