Support read-only lease in NoteDb meta graph
To support reverse migration from NoteDb -> ReviewDb, we need to do a
similar thing to what we did with NoteDbChangeState, except in this case
NoteDb is the source of truth, so we need to record the lease in NoteDb.

Let a determined caller clear a lease by setting the readOnlyUntil to
the special timestamp of 0. There is no assurance that the caller is the
same one that holds the lease, but this code is going to be used in
exactly one place.

This does pollute the meta graph with implementation details of the
migration, which we didn't really want to do. We can erase our tracks
later by rebuilding the change again once ReviewDb is primary again.

Change-Id: Ia019fb47e1c1edf719354d0da6bbfe22429f3eb1

Former-commit-id: cfb8e840ae23ca5842c825c21250ac7ecd41d85b