Only count 'unresolved' status on leaf comments
The previous change [1] counts the 'unresolved' status on all the
comments of a change. However, currently when posting a new comment
to resolve a comment thread, the 'unresolved' status of its ancestors
will not be updated. But by design, we hope a comment thread to be
resolved if its leaf comment is resolved.

This change doesn't update RobotCommentsIT as RobotComment extends
the 'unresolved' and 'inReplyTo' fields from Comment. CommentIT
should be enough to cover this functionality.

[1] I97c6aeb47db48f56f94c5d184dfa36c8d7868cc8

Change-Id: I8253226d883f36ca10763caccf9af3fce0c29290

Former-commit-id: 056d6921de9d0954a48ea52e103e84e615668dd1