Use GR-TOOLTIP in GR-DATE-FORMATTER
The date formatter shows detailed date-timestamps in a native browser
tooltip. This feature has a low discoverability because of how long
users must hover before the native tooltip appears.

With this change, the date formatter uses PolyGerrit's tooltip behavior
so that the tooltip appears immediately and is easier to access.

Because message timestamps appear on the far right of the window, the
tooltip behavior is upgraded to offset the arrow to the right when
needed to align with their referents.

Change-Id: I6e9b13e6dab62b6df1e3a9114f7559185cc3bc70

Former-commit-id: 72821cbc5754eca6f7a87c99f6cb5acca6a61111