Change capabilities collection to parse using PermissionBackend
Add GlobalOrPluginPermission interface to allow PluginPermission to
represent a capability contributed by a plugin.  This can now be
tested through PermissionBackend instead of using CapabilityControl.

Parsing a capability uses test() to determine if the permission exists
for the target user.  Using test here is more correct than check as
the code isn't exercising the permission, its merely testing if the
user might be able to exercise the permission.

Update the capabilities REST API view to use the new PluginPermission,
which allows combining everything into a single test invocation.

Change-Id: I23f44ae3398a496fcb13421344ae545ac68cce93

Former-commit-id: e9e1af205c3927f6374628b3fbbe6f052be03c70