ApprovalsUtil: Throw RestApiExceptions from check* methods
This means callers don't have to catch the unchecked
IllegalArgumentException, which it's hard to guarantee is from this
block of code, and is easy to forget.

In particular, ChangeInserter previously let IAE propagate, which turned
into a generic error during ReceiveCommits if trying to set a disallowed
label. (Actually it's still a fairly generic error, but it's now at
least possible for future cleanup to ReceiveCommits to report the error
properly should it so desire.)

Use BadRequestException and AuthException to match the behavior of
PostReview for these cases.

Change-Id: I267e7168e254d6fc3fb54516feee089d69265d67

Former-commit-id: 4e6ffddaad83e8240eb8c4120d04f4f228375e02