InlineEdit: Trigger reindexing on git push
When pushing with magic branch %edit option, there was no operation
added and as a consequence no reindexing took place, which lead to
non working has:edit predicate.

To rectify, add dummy change edit mutation operation, that sets pseudo
dirty flag to trigger the reindexing. Update tests correspondingly to
verify that after the git push to refs/for/master%edit the index is
aware of the edit.

Reported-By: Alice Kober-Sotzek <aliceks@google.com>
Bug: Issue 5350
Change-Id: I4f366bcbb9d2aa45c288c636b1ba2c3e485f8fb4

Former-commit-id: 3cb7c81a68db130b3a8e6fd43c1bbc06f85e3359