VelocityRuntimeProvider: Change logger channel name back to 'velocity'
In change Ie126ed263 the logger channel name was changed to use the
class name.  A side-effect of that is that the logger configuration
in log4j.properties:

  log4j.logger.velocity=WARN

is now ineffective.

Change the channel name back to its previous value, so that the log4j
configuration works again.

We could update the configuration to match the new channel name, but
this would also require users to make the same change in their own
configurations.

Change-Id: I363c9674aefafa105f3f96078ff8a9dc71f698ee

Former-commit-id: 646db67977a812354abf645b3dda290180de339e