Limit diff against to values larger than base
Previously, there was no check as to whether the patchNumber was larger
than the basePatch. When a user selected values that did not make sense,
the diffs did not reload.

This change addresses the problem by limiting the patch number selection
to values that are larger than the current base patch number.

The base patch number is passed back to the change view via a 2 way data
binding of the patchRange.

Bug: Issue 5613
Change-Id: Iec52890414dfd56e6c8070306fd39a9567627a47

Former-commit-id: 5da21a00cb8c053f6633329ec1ddc214a0645e22