Support "cc:" query
For NoteDb, where REVIEWER/CC are truly distinct, change "reviewer:" to
only query for REVIEWER. Keep the old behavior of "reviewer:" for
non-NoteDb and return results regardless of what state it guessed. For
"cc:", actually search for CC, although in practice usually this will
return no results (matching the fact that CCs don't show up in the UI).

Change-Id: I0b620888f93d64cea1d2e34cefb8d47ee026f0fd

Former-commit-id: 49267dd816293836447b3158ce06d28eeb128d9d