Fix next/prev page event handlers on change list view
There was a logic flaw that prevented "]" or "n" from working to
navigate to the next page with an offset greater than the items per
page, resulting in the keyboard shortcuts only working the first time.

This change fixes the issue and also adds some test coverage to this
element.

Bug: Issue 5537
Change-Id: I8893a8517eeb05c5228258b8efa3090bc076057e

Former-commit-id: d405bbaca3af7cbcfb62b8a329456889cf36e73f