Don't show next arrow when there are not more changes on next page
The logic hiding the next arrow on the change list view was incorrect
when there was a full page of results, but no more to dipslay.

The API includes an attribute _more_changes on the last item of the
array to let the client know whether or not there should be another
page. This change uses _more_changes instead of comparing to
_changesPerPage to take this edge case into account.

Bug: Issue 5622
Change-Id: I17dba2c807d152485494405fb412a502ba666df6

Former-commit-id: 4491ff19a476ae68ad0b87b5c3c959bfdba17f8c