Modify change table editor so that all columns are in a single section
Previously the change table editor had a section for shown columns and
a separate one for hidden columns. From the shown columns section, you
could 'delete' a column, and from the hidden section, you could 'add'
a column. This was confusing for users, particularly when there were no
hidden columns.

This change keeps all columns in the section, but instead adds checkbox
indicating whether or not each column should be visible.

Bug: Issue 5547
Change-Id: I65d1afd4c5e3e9cbbf823ad5773db1f13ca182ef

Former-commit-id: 14aa6ab6fccc85e775fb6a4aa5597b172e504281