Less related changes jank
WHen navigating between changes, especially large changes or changes
with large "related changes" sections, the relation chain for the
previous change may be briefly shown before reloading and reflowing the
change view below.

With this change, the "related changes" section is put into it's
"unloaded" state when the change view reloads so that it will not
display outdated entries and will reflow the page *fewer* times.

Bug: Issue 5627
Change-Id: I8904e2bb387c0386ceec09d59413858b8038fd9b

Former-commit-id: 98dae525bc3ee64a8017cc3efc6cd933234eb311