Hoist validation from SetAssigneeOp into PutAssignee
Performing validation of the new asignee inside the operation isn't
necessary; it can be validated in the REST API handler as its parsing
the input object and preparing the operation object.  This
signficantly cleans up the operation object, making it easier to
follow the logic.

This change also makes it easier to later refactor the check "can new
assignee see the target branch", before the operation is even started.

Change-Id: I3c8f3459300a670e8c76edff1262399de704933b

Former-commit-id: 8ca95f4e25085f0ce2026adc93d3db8ba4300725