Convert PutAssignee to PermissionBackend
Use check(ChangePermission.READ) to ensure the assignee can
read the change. It's assumed crazy to assign a change to a
user that is not able to read the change. For a draft change
this may require adding them to the reviewer list first to
ensure they have the necessary read permission.

Change-Id: I4fc278c6bdf739bb603b2c8bd490f3480ba99663

Former-commit-id: 88200f3907ddff6be0a38ee56b9fbaf77c6fff00