Do not include cookbook plugin in release.war
3005bf6afed5c1d33b6de97dac85407d620ca32d [formerly 1164e53283187b8a2a79052f5ea85471f6f0b6d4] (Allow to add custom core
plugins in release build, 2015-12-11) added an item CUSTOM to the buck
plugin rules to make it easier for people to add additional 'core'
plugins in a release build of a forked copy of Gerrit.

When that was ported to bazel, it included cookbook-plugin by mistake.
As a result, when someone runs

  bazel build release

they get a copy of gerrit with the cookbook plugin included and
enabled, which is almost never what they wanted.

Fix it by making the default list of custom plugins an empty list again.

Change-Id: I9b1653046bd368b8ec5c1978a15bd675c5dd31e7

Former-commit-id: 021d7df24c42fb78222acbbf01606b91946c7d65