Enhance plugin extension panels to allow ordering per a config
Enhance plugin extension panels to allow ordering to be specified
in AllProjects projects config so that ordering can be defined rather
than determined by the order in which plugins are loaded.

The interface change is backwards compatible at runtime
because the extension panel implementation will correctly
handle the null value for the panel name.

Change-Id: I51f3ac5a18b504daa5bf55d92cc33621d2247082

Former-commit-id: e7f3d1adfc43d201745ca2fbd77f2d5b106499c4