Fix problem where gr-formatted-text would sometimes not display
There was an issue where gr-formatted-text would not display when the
config was not loaded yet.  This change adds a function to display text
as is, in the event that the config is not yet loaded. It also refactors
the existing functions to make it more clear where config is needed.

Bug: Issue 5690
Change-Id: I74896bd59793b26b2b8fe289c13e5762c60fe8df

Former-commit-id: e76c79bc055aa61d2fceb128b0750885ed466faa