Allow user with maintain server permissions to find any change
Gerrit users who are delegated to maintain the server may be highly
restricted to access projects and branches.
There are maintenance operations that can still be allowed
on change-ids even on non-visible projects.

By allowing users delegated to maintain the server to find
non-visible change-ids in the CLI we can enable features such as
the ad-hoc change indexing.

Change-Id: Ia2ec07517c70f2cc4aed1bee832ea04a45fc6466

Former-commit-id: fe6f7395a2bac2fa33da2219714c8e5f0a089de1