Fix race in editing account info
Because modifying the name and status field in account-info requires two
separate API calls, a race was occurring server side that caused the DB
to reject the modification of the status field most of the time.

Before, the requests to set status and name were made every time. Now,
the requests are made synchronously, and are only made if the
corresponding property of the account object was actually modified.

Bug: Issue 5721
Change-Id: I38d2927cbd0998a0debc414ed09ca1f74e8354ff

Former-commit-id: 5827312720d39d97b42cf20d5567f9ea7dec5c51