BatchUpdate: Extract interfaces for *Context
Keep the current implementations, which depend on reaching into
BatchUpdate internal state. This results in parallel inheritance for the
interfaces and implementations, which is a little sketchy and we might
eventually get rid of. But it keeps the change pretty minimal for now.

While we're extracting new public interfaces, document them.

Change-Id: I4f9ec67b1a2a560428aef01e41a3a94cb6ce6912

Former-commit-id: b91cf223be6ba9fad3c1757397d4190c8b2b817e