Incremental rendering for large shared groups under full-context
Formerly, incremental rendering had not been applied to shared chunks
because it would break context-grouping logic. However, when the context
is set to the whole file, this problem is averted because the
context-groups are not created. Apply incremental rendering to shared
chunks only when the context is set to whole file.

Change-Id: I8018535ee72e5d4ef499628181298a13e8b20d06

Former-commit-id: 1ce7fadaef6ed42661e7294485a387b6bb71b6d6