Allow fixes to refer to several files
Suggested fixes may now also refer to other files than the one on which
the robot comment was placed. In addition, a fix may modify several
files.

Whether a file exists or not is only verified when a fix is applied.
The check isn't done when a robot comment is added in order to not
slow down the operation.

Change-Id: Id26232019b42dfe8286c9b0696f00b8a376482b0

Former-commit-id: 791f4af42ce6b3ce635c19b612faa1bdd73ee71a