Set file status to reviewed when expanded inline
Modifies the state of the checkbox and makes an API call when a file
is expanded inline.

As part of this change, the logic for determining whether or not a
reviewed checkbox should be checked is refactored from a computed
function to a computed property set on the file itself.

Also fixes a style error in the diff-builder.

Bug: Issue 5798
Change-Id: Ib1297d6c6d7e000e67465c20f21913063d360ff1

Former-commit-id: 2700a4d74ec798530de8825e40805c5e56b058d4