Discard request HTTP bodies before writing response
The HTTP RFCs require a server to fully consume the request body before
it can return a non-error status code.

Servlet API allows to use either InputStream or Reader. The code changes
switches the way to discard based on the request type.

There is a case that it discards the body even if it has an error status
code. The performance impact from this would be negligible because this
happens only when the REST API takes a raw input and it returns a
Response object with an error status code instead of throwing an
exception.

The same has been done in JGit in https://git.eclipse.org/r/#/c/4716/.

Change-Id: I3a40cbaf2c89e158451a7ed7c06fc98744d87f52

Former-commit-id: 8f96ddd16f0e6b6b84c29cb13e643c477a95b54c