Fresh NoteDb implementation of BatchUpdate
This implementation doesn't have to deal with dual ReviewDb/NoteDb
setups, so its implementation of updating changes is much simpler.
Use the new implementation when noteDb.changes.disableReviewDb=true.

Large portions of this new code are copy/pasted from the
ReviewDbBatchUpdate implementation. Minimizing code duplication is a
non-goal in this change; soon we will more heavily modify this
implementation to fuse the repo and change update phases, so it's not
clear how much duplication there will be in the end.

Change-Id: I9eefdb1268685515449ef15842d15f50bdcbc7bd

Former-commit-id: fb4aeecf05df27a92f32f0ba70d569f4c4840fff