Test stale credentials on visibility change
When returning to a page after a long time, test the credentials
immediately after the page has been foregrounded and show the auth-toast
if they are found to be expired.

Currently, while the user is expected to be refreshing their
credentials, the GR-ERROR-MANAGER listened to `visibilityChange` events
to hasten authentication checks. With this change, the error manager
always listens to `visibilityChange` events (rather than only when
refreshing). In response to these events it will:
- Check the credential refresh status (if the user is expected to be
  refreshing credentials).
- Validate the user's credentials (if the credentials have not been
  validated recently).
- Otherwise do nothing.

Change-Id: Ic9c0a739043cf7b35b89219739fd9bc5dba51634

Former-commit-id: 516d08a4bd106ff95c865c626d7ff1eae252f3de