Restrict size of a robot comment
Robot comments whose JSON representation exceeds the specified amount
of bytes are rejected on addition. This precautionary measure is
introduced because those comments are typically generated by tools
which might easily generate a huge comment erroneously.

The limit can be specified in the gerrit.config. The default limit is
1024kB. Zero or negative values allow robot comments to be of unlimited
size.

Change-Id: I75b3dfd02266d4223cc0c7e46d3b3f1d8c8b0572

Former-commit-id: 5ccf7078ab879b100ae36983afb98b847d2d6f17